Ajustes de permisos para mostrar botones de menu

+++ database/migrations/2024_11_27_162155_create_inventarios_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->integer('StockApertura');
             $table->integer('ProductosCierre');
             $table->integer('StockCierre');
-            $table->float('totalInventario');
+            $table->decimal('totalInventario');
             $table->foreignId('aperturadoPor')->references('id')->on('users');
             $table->foreignId('cerradoPor')->references('id')->on('users');
             $table->foreignId('estado')->references('id')->on('estados');
